requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LLM integration
anthropic>=0.25.0
openai>=1.0.0

# Data download
kagglehub>=0.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
